Add example for loading data from a file
Previously it was hard coded, and many examples hard code the data in
the javascript, which is not what we want for visualizing an external
dataset.

This shows a simple working example of loading the data from disk.

Tested using:

```
python3 -m http.server
```

You should see a blue shape in the browser, next to "Polygon Hull:".